remove .git directories

+++ test/install-deps.nix
@@ -5,8 +5,10 @@ let
   packageDrvs = builtins.attrValues packages.inputs;
 
   copyCmds = map (x: let target = ".psc-package/${packages.set}/${x.name}/${x.version}"; in ''
-    mkdir -p ${target}
-    cp --no-preserve=mode,ownership,timestamp -r ${toString x.outPath}/* ${target}
+    if [ ! -e ${target} ]; then
+      mkdir -p ${target}
+      cp --no-preserve=mode,ownership,timestamp -r ${toString x.outPath}/* ${target}
+    fi
   '') packageDrvs;
 
 in pkgs.stdenv.mkDerivation {
